refactor: split AvatarGroup into own files
