+++ src/avatar/avatar.mdx
@@ -10,6 +10,7 @@ import {
 } from '@storybook/addon-docs/blocks'
 
 import * as AvatarStories from './avatar.stories'
+import * as AvatarGroupStories from './avatar-group.stories'
 
 <Meta of={AvatarStories} />
 
@@ -34,7 +35,7 @@ hidden from assistive tech. When the group represents a labeled entity (a
 button, link, or labeled region), supply an `aria-label` on the group that
 conveys the count to assistive tech.
 
-<Canvas of={AvatarStories.AvatarGroups} />
+<Canvas of={AvatarGroupStories.People} />
 
 ## Migrating from the legacy API
