+++ src/avatar/index.ts
@@ -1,1 +1,2 @@
 export * from './avatar'
+export * from './avatar-group'
